requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
chromadb==0.4.22
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
doc2docx==0.2.4
numpy==1.24.3